+++ script/sccache.sh
@@ -16,13 +16,20 @@ LATEST_URL=`github_latest_url "$LATEST_URL"`
 LATEST_TAG=`echo $LATEST_URL | awk -F '/' '{print $NF}'`
 [ -f "$SAVE_PATH/$SAVE_DIR/.$LATEST_TAG" ] && echo "Installation successful." && exit 0
 REMOTE_FILE_PREFIX="sccache-${LATEST_TAG}-x86_64-unknown-linux-musl"
-download "$SAVE_PATH/$SAVE_FILE_PREFIX$SAVE_FILE_SUFFIX" "$LATEST_URL/$REMOTE_FILE_PREFIX$SAVE_FILE_SUFFIX"
 
 backup $SAVE_PATH/$SAVE_DIR
+cd $SAVE_PATH
 
-cd $SAVE_PATH && wrap_decompress $SAVE_DIR $SAVE_FILE_PREFIX$SAVE_FILE_SUFFIX
+if [ -n "$SAVE_FILE_SUFFIX" ]; then
+    download "$SAVE_PATH/$SAVE_FILE_PREFIX$SAVE_FILE_SUFFIX" "$LATEST_URL/$REMOTE_FILE_PREFIX$SAVE_FILE_SUFFIX"
+    wrap_decompress $SAVE_DIR $SAVE_FILE_PREFIX$SAVE_FILE_SUFFIX
+    rm -f $SAVE_PATH/$SAVE_FILE_PREFIX$SAVE_FILE_SUFFIX
+else
+    mkdir -p $SAVE_PATH/$SAVE_DIR
+    download "$SAVE_PATH/$SAVE_DIR/$SAVE_FILE_PREFIX" "$LATEST_URL/$REMOTE_FILE_PREFIX"
+fi
+
 circulate_ln "$BIN_PATH" "${BIN[*]}" "$LOCAL_BIN_PATH"
 touch $SAVE_DIR/.$LATEST_TAG
-rm -f $SAVE_PATH/$SAVE_FILE_PREFIX$SAVE_FILE_SUFFIX
 echo "Installation successful."
